Fix off-by-one bug in poll/batch/offset mgmt
[Card](https://rentpath.atlassian.net/browse/SRV-1287)

The first record of the second and later batches was being skipped due
to the way we were building the source record offset based on the reader
offset after reading a record.

An aside: seems weird to me that the reader offset is 1
(not 0) before the first read. Seems to me offsets should be 0-based.